# cali_lights_openapi.yaml
openapi: 3.0.3
info:
  title: Cali Lights API
  version: "1.0.0"
  description: >
    Networked Triad Mission platform: chains, invites, missions, gallery, metadata analysis,
    Veo video generation, and invite/bridge flows. Every endpoint maps directly to checklist B.
servers:
  - url: https://api.cali-lights.example
...
security:
  - bearerAuth: []
paths:
...
  /api/invite/create:
    post:
      tags: [Invites]
      summary: Generate an invite token and QR for a chain
      operationId: createInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chainId: { type: string, format: uuid }
                expiresInHours: { type: integer, minimum: 1, maximum: 168, default: 72 }
              required: [chainId]
      responses:
        "200":
          description: Invite created
          content:
            application/json:
              schema:
                type: object
                properties:
                  invite: { $ref: '#/components/schemas/Invite' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /api/mission/lock:
    post:
      tags: [Missions]
      summary: Transition a mission into FUSING state when capture window closes
      operationId: lockMission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                missionId: { type: string, format: uuid }
                reason: { type: string, enum: [timer, full, manual], default: timer }
              required: [missionId]
      responses:
        "200":
          description: Mission locked for fusing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Mission' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/mission/recap:
    post:
      tags: [Missions]
      summary: Run recap pipeline, kicking off metadata blending and Veo request
      operationId: recapMission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                missionId: { type: string, format: uuid }
                force: { type: boolean, default: false }
              required: [missionId]
      responses:
        "200":
          description: Recap pipeline started
          content:
            application/json:
              schema:
                type: object
                properties:
                  chapter: { $ref: '#/components/schemas/Chapter' }
                  videoOperation: { $ref: '#/components/schemas/VideoOperation' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "409": { $ref: '#/components/responses/Conflict' }
  /api/mission/archive:
    post:
      tags: [Missions]
      summary: Archive a mission after recap and update streaks
      operationId: archiveMission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                missionId: { type: string, format: uuid }
                publishAt: { type: string, format: date-time, nullable: true }
              required: [missionId]
      responses:
        "200":
          description: Mission archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  chapter: { $ref: '#/components/schemas/Chapter' }
                  streaks:
                    type: array
                    items:
                      type: object
                      properties:
                        userId: { type: string, format: uuid }
                        currentStreak: { type: integer }
                        longestStreak: { type: integer }
        "401": { $ref: '#/components/responses/Unauthorized' }
...
  /api/metadata/analyse:
    post:
      tags: [Entries]
      summary: Analyse media and extract metadata
      operationId: analyseMetadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entryId: { type: string, format: uuid }
                mediaUrl: { type: string, format: uri }
                mediaType: { type: string, enum: [photo, video] }
                chainId: { type: string, format: uuid }
                userId: { type: string, format: uuid }
              required: [entryId, mediaUrl, mediaType, chainId, userId]
      responses:
        "200":
          description: Metadata extracted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MetadataAnalysis' }
        "401": { $ref: '#/components/responses/Unauthorized' }
...
  /api/video/status:
    get:
      tags: [Video]
      summary: Get status of a Veo generation operation
      operationId: getVideoStatus
      parameters:
        - in: query
          name: operationId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Operation status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideoStatus' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict with current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Invite:
      type: object
      properties:
        id: { type: string, format: uuid }
        chainId: { type: string, format: uuid }
        token: { type: string }
        inviteUrl: { type: string, format: uri }
        qrSvgUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
        acceptedAt: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [pending, redeemed, expired] }
        createdBy: { type: string, format: uuid }
...
    Mission:
      type: object
      properties:
        id: { type: string, format: uuid }
        chainId: { type: string, format: uuid }
        prompt: { type: string }
        windowSeconds: { type: integer }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        state:
          type: string
          enum: [IDLE, LOBBY, CAPTURE, FUSING, RECAP, ARCHIVED]
        submissionsRequired: { type: integer, default: 3 }
        submissionsReceived: { type: integer, default: 0 }
        presenceCount: { type: integer, default: 0 }
...
    Chapter:
      type: object
      properties:
        id: { type: string, format: uuid }
        missionId: { type: string, format: uuid }
        title: { type: string }
        poem: { type: string }
        collageUrl: { type: string, format: uri, nullable: true }
        videoUrl: { type: string, format: uri, nullable: true }
        soundtrackUrl: { type: string, format: uri, nullable: true }
        durationSeconds: { type: integer, nullable: true }
        generatedAt: { type: string, format: date-time, nullable: true }
        finalPalette:
          type: array
          items: { type: string }
        isShareable: { type: boolean, default: false }
        shareUrl: { type: string, format: uri, nullable: true }
        shareExpiresAt: { type: string, format: date-time, nullable: true }
    MetadataAnalysis:
      type: object
      properties:
        entryId: { type: string, format: uuid }
        dominantHue: { type: integer }
        palette:
          type: array
          items: { type: string }
        sceneTags:
          type: array
          items: { type: string }
        objectTags:
          type: array
          items: { type: string }
        motionScore: { type: number }
        gpsCity: { type: string }
        analysisTimestamp: { type: string, format: date-time }
...
    VideoStatus:
      type: object
      properties:
        operationId: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, SUCCEEDED, FAILED] }
        videoUrl: { type: string, format: uri, nullable: true }
        durationSeconds: { type: integer, nullable: true }
        watermark: { type: string, nullable: true }
        costUsd: { type: number, format: float, nullable: true }
        generatedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
    VideoOperation:
      type: object
      properties:
        id: { type: string, format: uuid }
        operationId: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, SUCCEEDED, FAILED] }
        targetType: { type: string, enum: [chapter, invite, bridge] }
        targetId: { type: string, format: uuid }
        prompt: { type: string }
        inputMediaUrls:
          type: array
          items: { type: string, format: uri }
        aspectRatio: { type: string }
        lengthSeconds: { type: integer }
        model: { type: string }
        costUsd: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
...